requirements.txt
altgraph==0.17.2
appdirs==1.4.4
async-generator==1.10
attrs==21.4.0
auto-py-to-exe==2.17.0
beautifulsoup4==4.10.0
bottle==0.12.19
bottle-websocket==0.2.9
bs4==0.0.1
Cartopy @ file:///C:/Users/Hanzla/Downloads/Cartopy-0.20.2-cp37-cp37m-win_amd64.whl
cattrs==1.10.0
certifi==2021.10.8
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.12
click==8.0.3
colorama==0.4.4
comtypes==1.1.11
cryptography==36.0.1
cx-Freeze==6.10
cx-Logging==3.0
cycler==0.11.0
decorator==5.1.1
Eel==0.12.4
enum34==1.1.10
Flask==2.0.3
fonttools==4.29.1
future==0.18.2
geocoder==1.38.1
geographiclib==1.52
geopy==2.2.0
gevent==21.12.0
gevent-websocket==0.10.1
googletrans==3.0.0
goslate==1.5.2
greenlet==1.1.2
gTTS==2.2.3
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2021.12.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
importlib-metadata==4.11.1
itsdangerous==2.0.1
jaraco.context==4.1.1
Jinja2==3.0.3
joblib==1.1.0
keyboard==0.13.5
kiwisolver==1.3.2
lief==0.11.5
MarkupSafe==2.0.1
matplotlib==3.5.1
more-itertools==8.12.0
MouseInfo==0.1.3
nltk==3.7
numpy==1.21.5
opencv-python==4.5.5.62
outcome==1.1.0
packaging==21.3
pefile==2021.9.3
Pillow==9.0.1
playsound==1.2.2
psutil==5.9.0
py-cpuinfo==8.0.0
PyAudio @ file:///D:/jarvis/PyAudio-0.2.11-cp37-cp37m-win_amd64.whl
PyAutoGUI==0.9.53
pycaw==20181226
pycparser==2.21
PyDictionary==2.0.1
pydub==0.25.1
PyGetWindow==0.0.9
pyinstaller==4.9
pyinstaller-hooks-contrib==2022.2
pyjokes==0.6.0
PyMsgBox==1.0.9
pyOpenSSL==22.0.0
pyparsing==3.0.7
PyPDF2==1.26.0
pyperclip==1.8.2
pypiwin32==223
pyproj==3.2.1
PyQt5==5.15.6
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.1
PyRect==0.1.4
PyScreeze==0.1.28
pyshp==2.2.0
PySocks==1.7.1
python-dateutil==2.8.2
python-decouple==3.6
pyttsx3==2.90
pytube==12.0.0
pytweening==1.0.4
pywhatkit==5.3
pywikihow==0.5.7
pywin32==303
pywin32-ctypes==0.2.0
ratelim==0.1.6
regex==2022.1.18
requests==2.27.1
requests-cache==0.9.1
requests-futures==1.0.0
rfc3986==1.5.0
screen-brightness-control==0.11.5
selenium==4.1.0
Shapely==1.8.1.post1
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.1
SpeechRecognition==3.8.1
speedtest-cli==2.1.3
stem==1.8.0
torch==1.10.2
torrequest==0.1.0
tqdm==4.62.3
trio==0.19.0
trio-websocket==0.9.2
typing_extensions==4.1.1
url-normalize==1.4.3
urllib3==1.26.8
Werkzeug==2.0.3
whichcraft==0.6.1
wikipedia==1.4.0
winshell==0.6
WMI==1.5.1
wolframalpha==5.0.0
wsproto==1.0.0
xmltodict==0.12.0
zipp==3.7.0
zope.event==4.5.0
zope.interface==5.4.0
